% tinyrox says don't edit this manually, but it can't stop you!
\name{nloptVersion}
\alias{nloptVersion}
\title{NLopt Version as Vector}
\usage{
nloptVersion()
}
\value{
A list with an integer vector with three elements for major, minor
and patch release, classed as \code{package_version} and \code{numeric_version}.
}
\description{
NLopt Version as Vector
}
\examples{
nloptVersion()
}
