% tinyrox says don't edit this manually, but it can't stop you!
\name{wrapper}
\alias{wrapper}
\title{Minimal wrapper function}
\usage{
wrapper(method = "MMA")
}
\arguments{
\item{method}{A string defaulting to \sQuote{MMA} (also allowing \sQuote{COBYLA})
which selects the algorithm use.}
}
\value{
Nothing is returned as the function is invoked for
its side effects.
}
\description{
Minimal wrapper for C++ function to illustrate
logging from R
}
