% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nloptVersion}
\alias{nloptVersion}
\title{NLopt Version as Vector}
\usage{
nloptVersion()
}
\value{
Am integer vector with three elements for major, minor and patch release.
}
\description{
Helper function to access the NLopt version as an integer vector.
}
\examples{
nloptVersion()
package_version(paste(as.character(nloptVersion()), collapse="."))
}
